deleted unneeded files and made it responsive
