--- src/containers/ForgetPassword/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ForgotPassword from './forgotPass';
-
-function index() {
-  return <ForgotPassword />;
-}
-
-export default index;
