+++ src/containers/ForgotPassword/forgotPass.css
@@ -1,0 +1,101 @@
+.forgot-password {
+  min-height: 100vh;
+}
+
+.forgot-password__heading {
+  display: flex;
+  justify-content: center;
+  align-items: baseline;
+  font-family: 'VAGRounded', sans-serif;
+}
+
+.forgot-password__back-btn {
+    flex: 1;
+}
+
+.forgot-password__title {
+    flex: 6;
+    text-transform: capitalize; 
+}
+
+.forgot-password__vector {
+    margin: 10px auto;
+}
+
+.forgot-password__text {
+  width: 80%;
+  margin: 20px auto;
+  font-size: 14px;
+  color: #242424;
+  font-family: 'Open Sans', sans-serif;
+  font-weight: 300;
+  text-align: left;
+}
+.forgot-password__icon {
+    border: 1px solid #d9d9d9;
+    border-radius: 4px;
+    width: 15%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: #242424;
+}
+
+.forgot-password__resend {
+  margin-left: 4px;
+  color: #e9663a;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.forgot-password__input {
+  width: 80%;
+  margin: 10px auto;
+  display: flex;
+}
+
+.forgot-password .ant-btn {
+  border: none;
+}
+
+.forgot-password .ant-btn-primary {
+  color: #fff;
+  background-color: #e9663a;
+  font-weight: 600;
+  border: none;
+  text-shadow: none;
+  box-shadow: none;
+}
+
+.forgot-password .ant-btn:hover,
+.forgot-password .ant-btn:focus {
+  color: #fff;
+  background-color: #e9663aec;
+  border-color: #e9663aec;
+}
+
+.forgot-password .ant-btn-primary:hover,
+.forgot-password .ant-btn-primary:focus {
+  color: #fff;
+  background-color: #e9663aec;
+  border-color: #e9663aec;
+}
+
+.forgot-password__button {
+  width: 80%;
+  margin: 20px 0;
+}
+
+@media (min-width: 600px) {
+  .forgot-password__button,
+  .forgot-password__input,
+  .forgot-password__text {
+    width: 60%;
+  }
+  .forgot-password__text {
+    font-size: 20px;
+  }
+  .forgot-password__heading  {
+    font-size: 24px;
+  }
+}
